Chatbot/requirements.txt
sentence-transformers
qdrant-client
google-generativeai
python-dotenv
Pillow
pyvi